last

+++ database/migrations/2022_08_11_123851_create_applicants_table.php
@@ -16,8 +16,8 @@ class CreateApplicantsTable extends Migration
     {
         Schema::create('applicants', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('phone')->nullable();
-            $table->integer('pin')->nullable();
+            $table->bigInteger('phone')->nullable();
+            $table->bigInteger('pin')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
